+++ VGEngine/Engine/source/game/physicsSystem.cpp
@@ -47,7 +47,7 @@ void PhysicsSystem::update(std::vector<GameObject*> *gameObjects, float deltaTim
 
 		 if (physComponent != nullptr && transform != nullptr)
 		 {
-			 transform->setPosition(Vector2<int>((physComponent->getBody()->GetPosition().x  * scale) - transform->getSize().getX() / 2.0f, 
+			 transform->setPosition(Vector2<float>((physComponent->getBody()->GetPosition().x  * scale) - transform->getSize().getX() / 2.0f, 
 												(-physComponent->getBody()->GetPosition().y * scale) - transform->getSize().getY() / 2.0f));
 			  transform->setRotation(-1.0f * physComponent->getBody()->GetAngle() * 180 / 3.14);
